requirements.txt
Flask==3.1.0
numpy==2.2.5
Pillow==11.2.1
protobuf==6.30.2
python-dotenv==1.1.0
waitress==3.0.2
